# frontend/openapi/afisha.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Afisha API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/safety/reports": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Create Report",
        "operationId": "create_report_safety_reports_post",
        "parameters": [
          {
            "name": "X-Afisha-CSRF",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Afisha-Csrf"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "afisha_mini_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Afisha Mini Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Report Safety Reports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/account/cases": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "Cases Feed",
        "operationId": "cases_feed_account_cases_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "resolved"
              ],
              "type": "string",
              "default": "active",
              "title": "Status"
            }
          },
          {
            "name": "afisha_mini_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Afisha Mini Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cases Feed Account Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/account/cases/{case_public_id}": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "Case Detail",
        "operationId": "case_detail_account_cases__case_public_id__get",
        "parameters": [
          {
            "name": "case_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Public Id"
            }
          },
          {
            "name": "afisha_mini_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Afisha Mini Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Case Detail Account Cases  Case Public Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/account/cases/{case_public_id}/appeal": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Appeal Case",
        "operationId": "appeal_case_account_cases__case_public_id__appeal_post",
        "parameters": [
          {
            "name": "case_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Public Id"
            }
          },
          {
            "name": "X-Afisha-CSRF",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Afisha-Csrf"
            }
          },
          {
            "name": "afisha_mini_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Afisha Mini Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Appeal Case Account Cases  Case Public Id  Appeal Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppealBody": {
        "properties": {
          "explanation": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "explanation"
        ],
        "title": "AppealBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportBody": {
        "properties": {
          "subject_type": {
            "type": "string",
            "enum": [
              "event",
              "profile",
              "looking_post",
              "chat_message"
            ],
            "title": "Subject Type"
          },
          "subject_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Subject Id"
          },
          "reason_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9_]+$",
            "title": "Reason Code"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "subject_component": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "photo",
                  "title",
                  "description",
                  "avatar",
                  "background",
                  "bio",
                  "display_name",
                  "body",
                  "answer",
                  "message",
                  "whole"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Component"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "subject_type",
          "subject_id",
          "reason_code"
        ],
        "title": "ReportBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}